Generate counselor list PDF in the browser
PDFReactor is buggy and expensive and we should stop using it. Using a
library called jsPDF it's possible to generate PDFs in the browser.
Positioning elements is a little tricky but because we always return 10
results it's not so bad. The PDF generating script is loaded asynchronously
so front-end performance shouldn't be impacted.

https://github.com/MrRio/jsPDF