Copy landing page for complaintdatabase from complaint satellite app
The landing page (found at /data-research/consumer-complaints/) was the
only page that the complaint satellite app was still serving. With this
change, we can stop including complaint in our requirements and
deactivate the repository.

This update contains the view, templates, images, and static css
necessary to render the page.

TODO:
- Port any .less files necessary from the complaint app to style the
  page and generate static css.
- The page pulls in data from s3. Update the page to use dummy data
  instead, (only) in test environments
- Add any unit tests from the satellite app